+++ LoginNPushApp/server/app.js
@@ -2,6 +2,7 @@ const express = require('express');
 const axios = require('axios');
 const bodyParser = require('body-parser');
 
+//Create an instance of the express server
 const app = express();
 
 app.use(bodyParser.json());
@@ -11,14 +12,18 @@ app.get('/test', (req, res) => {
   res.send('Server is ready to send push notifications!');
 });
 
+//push notication
 app.post('/push', async (req, res) => {
   const currentTime = new Date().toString();
 
+//Setting the headers to call the OneSignal REST API
   const headers = {
     'Content-Type': 'application/json; charset=utf-8',
     Authorization: 'Basic MTY1NzY0MzItOTFkOS00ZDViLWIxN2ItYzY4ZWRlOWU1ODAz'
   };
 
+//Setting the body for the API, along with the device ID to whom
+//the notification must be sent.
   const body = {
     app_id: '8b7bbc52-fa66-4daa-92fc-ce57e3216402',
     contents: { en: currentTime },
@@ -29,6 +34,7 @@ app.post('/push', async (req, res) => {
     headers
   };
 
+//Call to the REST API with the headers and the body
   try {
     const response = await axios.post(
       'https://onesignal.com/api/v1/notifications',
